+++ apps/saas/app/layout.tsx
@@ -1,21 +1,8 @@
 import { config } from "@config";
-import { cn, Toaster } from "@repo/ui";
-import { ApiClientProvider } from "@shared/components/ApiClientProvider";
-import { ClientProviders } from "@shared/components/ClientProviders";
 import type { Metadata } from "next";
-import { NextIntlClientProvider } from "next-intl";
-import { getLocale, getMessages, getTranslations } from "next-intl/server";
-import { ThemeProvider } from "next-themes";
-import { Inter } from "next/font/google";
-
-import "./globals.css";
-import { NuqsAdapter } from "nuqs/adapters/next/app";
 import type { PropsWithChildren } from "react";
 
-const sansFont = Inter({
-	subsets: ["latin", "vietnamese"],
-	variable: "--font-inter",
-});
+import "./globals.css";
 
 export const metadata: Metadata = {
 	robots: {
@@ -28,34 +15,6 @@ export const metadata: Metadata = {
 	},
 };
 
-export default async function RootLayout({ children }: PropsWithChildren) {
-	const locale = await getLocale();
-	const messages = await getMessages();
-	const t = await getTranslations();
-
-	return (
-		<html lang={locale} suppressHydrationWarning className={sansFont.variable}>
-			<body className={cn("font-sans min-h-screen bg-background text-foreground antialiased")}>
-				<NuqsAdapter>
-					<NextIntlClientProvider messages={messages}>
-						<ThemeProvider
-							attribute="class"
-							disableTransitionOnChange
-							enableSystem
-							defaultTheme={config.defaultTheme}
-							themes={Array.from(config.enabledThemes)}
-						>
-							<ApiClientProvider>
-								<ClientProviders>
-									{children}
-
-									<Toaster position="top-right" closeLabel={t("common.aria.closeToast")} />
-								</ClientProviders>
-							</ApiClientProvider>
-						</ThemeProvider>
-					</NextIntlClientProvider>
-				</NuqsAdapter>
-			</body>
-		</html>
-	);
+export default function RootLayout({ children }: PropsWithChildren) {
+	return children;
 }